+++ cedar-policy-core/src/evaluator.rs
@@ -1926,12 +1926,11 @@ pub mod test {
                 ("bar".into(), Expr::val(4)),
                 ("foo".into(), Expr::val("hi")),
             ]),
-            Err(
-                expression_construction_errors::DuplicateKeyInRecordLiteralError {
-                    key: "foo".into()
-                }
-                .into()
-            )
+            Err(expression_construction_errors::DuplicateKeyError {
+                key: "foo".into(),
+                context: "in record literal",
+            }
+            .into())
         );
         // entity_with_attrs.address.street
         assert_eq!(
@@ -5533,12 +5532,11 @@ pub mod test {
         ]);
         assert_eq!(
             e,
-            Err(
-                expression_construction_errors::DuplicateKeyInRecordLiteralError {
-                    key: "a".into()
-                }
-                .into()
-            )
+            Err(expression_construction_errors::DuplicateKeyError {
+                key: "a".into(),
+                context: "in record literal",
+            }
+            .into())
         );
 
         let e = Expr::record([
@@ -5547,12 +5545,11 @@ pub mod test {
         ]);
         assert_eq!(
             e,
-            Err(
-                expression_construction_errors::DuplicateKeyInRecordLiteralError {
-                    key: "a".into()
-                }
-                .into()
-            )
+            Err(expression_construction_errors::DuplicateKeyError {
+                key: "a".into(),
+                context: "in record literal",
+            }
+            .into())
         );
 
         let e = Expr::record([